fix(backend): fix typo and improve card service
