+++ cards-backend/src/config/interfaces/services/card.service.interface.ts
@@ -8,4 +8,5 @@ import {CardsEntityInterface} from "../../../domain/cards/cards.entity.interface
 export interface CardServiceInterface {
 	getCards(params: any): Promise<CardsEntityInterface | CardsEntityInterface[] | null>;
 	refreshCardDatabase(): Promise<void>;
+	initializeCards(): Promise<void>;
 }
