+++ cards-backend/src/infrastructure/repositories/card.repository.ts
@@ -1,5 +1,3 @@
-import path from 'path';
-import fs from 'fs';
 import {DatabaseInterface} from "../../config/interfaces/adapters/database.interface";
 import {CardsEntityInterface, CardsEntityInterface as Card} from "../../domain/cards/cards.entity.interface";
 import {cardParameters, CardRepositoryInterface} from "../../config/interfaces/repositories/card.repository.interface";
@@ -15,17 +13,6 @@ export class CardRepository implements CardRepositoryInterface {
             return null;
         }
         return cards;
-    }
-
-    async initializeCards(): Promise<void> {
-        const filePath = path.resolve(__dirname, '../../src/scryfall/data/all_cards.json');
-        const rawData = fs.readFileSync(filePath, 'utf8');
-        const cards: Card[] = JSON.parse(rawData);
-        try {
-            await this.update(cards);
-        } catch (error) {
-            console.error("Error in initializeCards:", error);
-        }
     }
 
     async update(cards: Card[]): Promise<void> {
